+++ pyads/ads.py
@@ -6,6 +6,7 @@
 
 """
 import struct
+import itertools
 from collections import OrderedDict
 from ctypes import (
     memmove,
@@ -18,7 +19,7 @@ from ctypes import (
 )
 from datetime import datetime
 from functools import partial
-from typing import Optional, Union, Tuple, Any, Type, Callable, Dict, List, cast
+from typing import Optional, Union, Tuple, Any, Type, Callable, Dict, List, cast, Iterator
 
 # noinspection PyUnresolvedReferences
 from .constants import (
@@ -46,6 +47,7 @@ from .constants import (
     DATATYPE_MAP,
     ADSIGRP_SUMUP_READ,
     ADSIGRP_SUMUP_WRITE,
+    MAX_ADS_SUB_COMMANDS,
     ads_type_to_ctype,
 )
 from .filetimes import filetime_to_dt
@@ -850,15 +852,25 @@ class Connection(object):
         )
 
     def read_list_by_name(
-        self, data_names: List[str], cache_symbol_info: bool = True
+        self,
+        data_names: List[str],
+        cache_symbol_info: bool = True,
+        ads_sub_commands: int = MAX_ADS_SUB_COMMANDS,
     ) -> Dict[str, Any]:
-        """Read a list of variables in a single ADS call.
+        """Read a list of variables.
+
+        Will split the read into multiple ADS calls in chunks of ads_sub_commands by default.
+
+        MAX_ADS_SUB_COMMANDS comes from Beckhoff recommendation:
+        https://infosys.beckhoff.com/english.php?content=../content/1033/tc3_adsdll2/9007199379576075.html&id=9180083787138954512
 
         :param data_names: list of variable names to be read
         :type data_names: list[str]
         :param bool cache_symbol_info: when True, symbol info will be cached for future reading
-        :return adsSumRead: A dictionary containing variable names from data_names as keys and values read from PLC
-            for each variable
+        :param int ads_sub_commands: Max number of ADS-Sub commands used to read the variables in a single ADS call.
+            A larger number can be used but may jitter the PLC execution!
+
+        :return adsSumRead: A dictionary containing variable names from data_names as keys and values read from PLC for each variable
         :rtype dict(str, Any)
 
         """
@@ -874,20 +886,37 @@ class Connection(object):
                 i: adsGetSymbolInfo(self._port, self._adr, i) for i in data_names
             }
 
-        return adsSumRead(self._port, self._adr, data_names, data_symbols)
+        if len(data_names) <= ads_sub_commands:
+            return adsSumRead(self._port, self._adr, data_names, data_symbols)
+
+        return_data: Dict[str, Any] = {}
+        for data_names_slice in _list_slice_generator(data_names, ads_sub_commands):
+            return_data.update(
+                adsSumRead(self._port, self._adr, data_names_slice, data_symbols)
+            )
+        return return_data
 
     def write_list_by_name(
-        self, data_names_and_values: Dict[str, Any], cache_symbol_info: bool = True
+        self,
+        data_names_and_values: Dict[str, Any],
+        cache_symbol_info: bool = True,
+        ads_sub_commands: int = MAX_ADS_SUB_COMMANDS,
     ) -> Dict[str, ADSError]:
-        """Write a list of variables in a single ADS call
+        """Write a list of variables.
+
+        Will split the write into multiple ADS calls in chunks of ads_sub_commands by default.
+
+        MAX_ADS_SUB_COMMANDS comes from Beckhoff recommendation:
+        https://infosys.beckhoff.com/english.php?content=../content/1033/tc3_adsdll2/9007199379576075.html&id=9180083787138954512
 
         :param data_names_and_values: dictionary of variable names and their values to be written
         :type data_names_and_values: dict[str, Any]
         :param bool cache_symbol_info: when True, symbol info will be cached for future reading
+        :param int ads_sub_commands: Max number of ADS-Sub commands used to write the variables in a single ADS call.
+            A larger number can be used but may jitter the PLC execution!
 
-        :return adsSumWrite: A dictionary containing variable names from data_names as keys and values return codes
-            for each write operation from the PLC
-        :rtype dict(str, Any)
+        :return adsSumWrite: A dictionary containing variable names from data_names as keys and values return codes for each write operation from the PLC
+        :rtype dict(str, ADSError)
 
         """
         if cache_symbol_info:
@@ -909,7 +938,19 @@ class Connection(object):
                 for i in data_names_and_values.keys()
             }
 
-        return adsSumWrite(self._port, self._adr, data_names_and_values, data_symbols)
+        if len(data_names_and_values) <= ads_sub_commands:
+            return adsSumWrite(
+                self._port, self._adr, data_names_and_values, data_symbols
+            )
+
+        return_data: Dict[str, ADSError] = {}
+        for data_names_slice in _dict_slice_generator(
+            data_names_and_values, ads_sub_commands
+        ):
+            return_data.update(
+                adsSumWrite(self._port, self._adr, data_names_slice, data_symbols)
+            )
+        return return_data
 
     def read_structure_by_name(
         self,
@@ -1266,3 +1307,17 @@ class Connection(object):
             timestamp = filetime_to_dt(contents.nTimeStamp)
 
         return contents.hNotification, timestamp, value
+
+
+def _dict_slice_generator(dict_: Dict[Any, Any], size: int) -> Iterator[Dict[Any, Any]]:
+    """Generator for slicing a dictionary into parts of size long."""
+    it = iter(dict_)
+    for _ in range(0, len(dict_), size):
+        yield {i: dict_[i] for i in itertools.islice(it, size)}
+
+
+def _list_slice_generator(list_: List[Any], size: int) -> Iterator[List[Any]]:
+    """Generator for slicing a list into parts of size long."""
+    it = iter(list_)
+    for _ in range(0, len(list_), size):
+        yield [i for i in itertools.islice(it, size)]